# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/me/credential": {
      "get": {
        "description": "Returns the credential kind, its granted scopes and what those scopes actually amount to once the owner’s roles are applied. The MCP server filters its tool list by `effectiveScopes`.",
        "operationId": "current",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentCredentialResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "AUTH_001 / TOKEN_003 — No credential was presented, or it is invalid or expired",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "AUTH_002 / TOKEN_004 / TOKEN_005 / TOKEN_006 / TOKEN_007 — The caller's roles, or their credential's scopes, do not permit this",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Describe the calling credential and its effective permissions",
        "tags": ["API tokens"]
      }
    },
...
    "/api/v1/installations": {
      "get": {
        "description": "A read of the workspace’s own rows, not of GitHub. An empty list is what the console reads as “connect GitHub”, including on a deployment with no App configured.",
        "operationId": "list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InstallationResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "AUTH_001 / TOKEN_003 — No credential was presented, or it is invalid or expired",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "AUTH_002 / TOKEN_004 / TOKEN_005 / TOKEN_006 / TOKEN_007 — The caller's roles, or their credential's scopes, do not permit this",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the workspace’s connected GitHub installations",
        "tags": ["GitHub installations"]
      },
      "post": {
        "description": "Called with the `installation_id` and OAuth `code` GitHub puts on the install redirect. The code is exchanged once to prove the caller can see the installation, then discarded — it is never stored. Re-posting the same installation refreshes what GitHub reports about it.",
        "operationId": "connect",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectInstallationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstallationResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "GITHUB_005 / GITHUB_006 — The authorization code was expired or already used, or no organization is active",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "401": {
            "description": "AUTH_001 / TOKEN_003 — No credential was presented, or it is invalid or expired",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "GITHUB_004 — GitHub does not list that installation for the authorizing account",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "409": {
            "description": "GITHUB_003 — Another workspace already holds that installation",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "503": {
            "description": "GITHUB_002 — The GitHub App is not configured on this server",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Connect a GitHub App installation to the workspace",
        "tags": ["GitHub installations"]
      }
    },
    "/api/v1/installations/{id}/repositories": {
      "get": {
        "description": "Answered live by GitHub through the installation’s own token and cached for a minute. Nothing is mirrored: the installation is the allowlist and GitHub enforces it.",
        "operationId": "list",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RepositoryResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "AUTH_001 / TOKEN_003 — No credential was presented, or it is invalid or expired",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "AUTH_002 / TOKEN_004 / TOKEN_005 / TOKEN_006 / TOKEN_007 — The caller's roles, or their credential's scopes, do not permit this",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "GITHUB_001 — GitHub installation not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "409": {
            "description": "GITHUB_008 — The installation is suspended or no longer installed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "502": {
            "description": "GITHUB_009 — GitHub could not be reached or rejected the request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "503": {
            "description": "GITHUB_002 — The GitHub App is not configured on this server",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the repositories an installation covers",
        "tags": ["GitHub installations"]
      }
    },
...
  },
  "info": {
    "title": "Oppenheimer API",
    "description": "Oppenheimer REST API documentation",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "InvalidParamDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "email",
            "description": "Dotted path to the offending field"
          },
          "reason": {
            "type": "string",
            "example": "Invalid email",
            "description": "Why the value was rejected"
          }
        },
        "required": ["name", "reason"]
      },
      "ProblemDetailsDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "https://oppenheimer.dev/errors#user_001",
            "description": "URI reference identifying the problem type. `about:blank` when the status code says everything."
          },
          "title": {
            "type": "string",
            "example": "User not found",
            "description": "Short summary, stable per problem type"
          },
          "status": {
            "type": "number",
            "example": 404,
            "description": "HTTP status code"
          },
          "detail": {
            "type": "string",
            "example": "No user with id 3f1c…",
            "description": "Explanation specific to this occurrence"
          },
          "instance": {
            "type": "string",
            "example": "/api/v1/users/3f1c",
            "description": "URI reference of the specific occurrence — the request path"
          },
          "code": {
            "type": "string",
            "example": "USER_001",
            "description": "Machine-readable catalog code"
          },
          "correlationId": {
            "type": "string",
            "description": "Correlation id of the request, for log lookups"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-01-01T00:00:00.000Z",
            "description": "When it happened"
          },
          "invalidParams": {
            "description": "Per-field validation failures",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvalidParamDto"
            }
          }
        },
        "required": ["type", "title", "status"]
      },
...
      "CurrentCredentialResponseDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "description": "How the caller authenticated.",
            "enum": ["session", "api-token", "oauth"]
          },
          "userId": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "grantedScopes": {
            "type": "array",
            "description": "Scopes the credential carries. Null for a browser session, which is not scope-restricted.",
            "nullable": true,
            "items": {
              "type": "string",
              "enum": [
                "profile:read",
                "profile:write",
                "users:read",
                "users:write",
                "admin:read",
                "admin:write",
                "roles:read",
                "roles:write",
                "organizations:read",
                "organizations:write",
                "members:read",
                "members:write",
                "invitations:read",
                "invitations:write",
                "workspaces:read",
                "workspaces:write",
                "tokens:read",
                "tokens:write",
                "billing:read",
                "billing:write",
                "leads:read",
                "leads:write",
                "hosts:read",
                "hosts:write",
                "projects:read",
                "projects:write",
                "sessions:read",
                "sessions:write",
                "repositories:read",
                "repositories:write"
              ]
            }
          },
          "effectiveScopes": {
            "type": "array",
            "description": "What the credential can actually do: its scopes intersected with the owner’s current roles.",
            "items": {
              "type": "string",
              "enum": [
                "profile:read",
                "profile:write",
                "users:read",
                "users:write",
                "admin:read",
                "admin:write",
                "roles:read",
                "roles:write",
                "organizations:read",
                "organizations:write",
                "members:read",
                "members:write",
                "invitations:read",
                "invitations:write",
                "workspaces:read",
                "workspaces:write",
                "tokens:read",
                "tokens:write",
                "billing:read",
                "billing:write",
                "leads:read",
                "leads:write",
                "hosts:read",
                "hosts:write",
                "projects:read",
                "projects:write",
                "sessions:read",
                "sessions:write",
                "repositories:read",
                "repositories:write"
              ]
            }
          },
          "organizationIds": {
            "description": "Organizations the credential is restricted to, or null when unrestricted.",
            "nullable": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "kind",
          "userId",
          "email",
          "grantedScopes",
          "effectiveScopes",
          "organizationIds",
          "expiresAt"
        ]
      },
...
      "InstallationResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "The control-plane id. Everything but the connect body uses this."
          },
          "organizationId": {
            "type": "string",
            "format": "uuid"
          },
          "githubInstallationId": {
            "type": "number",
            "description": "GitHub’s own installation id.",
            "example": 45678901
          },
          "accountLogin": {
            "type": "string",
            "example": "acme-labs",
            "description": "The account the App is installed on."
          },
          "accountType": {
            "type": "string",
            "example": "Organization",
            "enum": ["User", "Organization"]
          },
          "repositorySelection": {
            "type": "string",
            "enum": ["all", "selected"],
            "description": "What the installation dialog granted. GitHub enforces it, not us."
          },
          "installedByUserId": {
            "type": "string",
            "format": "uuid",
            "description": "The account that connected it."
          },
          "suspendedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true,
            "description": "Set while GitHub reports the installation suspended; nothing resolves until it clears."
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "organizationId",
          "githubInstallationId",
          "accountLogin",
          "accountType",
          "repositorySelection",
          "installedByUserId",
          "createdAt",
          "updatedAt"
        ]
      },
      "ConnectInstallationRequest": {
        "type": "object",
        "properties": {
          "githubInstallationId": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true
          },
          "code": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": ["githubInstallationId", "code"]
      },
      "RepositoryResponseDto": {
        "type": "object",
        "properties": {
          "githubRepoId": {
            "type": "number",
            "description": "GitHub’s own repository id.",
            "example": 831004242
          },
          "name": {
            "type": "string",
            "example": "oppenheimer"
          },
          "fullName": {
            "type": "string",
            "example": "acme-labs/oppenheimer"
          },
          "defaultBranch": {
            "type": "string",
            "example": "main",
            "description": "Offered as the default base branch."
          },
          "private": {
            "type": "boolean"
          },
          "archived": {
            "type": "boolean",
            "description": "Archived on GitHub: readable, but pushes are refused."
          },
          "pushedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "When GitHub last saw a push, for ordering the picker."
          }
        },
        "required": [
          "githubRepoId",
          "name",
          "fullName",
          "defaultBranch",
          "private",
          "archived",
          "pushedAt"
        ]
      },
...
    }
  }
}