# apps/api/src/github/commands/connect-installation/connect-installation.request.dto.ts
import { connectInstallationSchema } from '@oppenheimer/shared';
import { createZodDto } from 'nestjs-zod';

export class ConnectInstallationRequest extends createZodDto(connectInstallationSchema) {}